services/orchest-api/requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
-e file:../../orchest-cli
    # via -r requirements-dev.in
-e file:../../lib/python/orchest-internals
    # via -r requirements-dev.in
aiohttp==3.7.4
    # via -r requirements-dev.in
alembic==1.7.7
    # via flask-migrate
amqp==5.0.9
    # via
    #   -r requirements-dev.in
    #   kombu
aniso8601==9.0.1
    # via flask-restx
apscheduler==3.7.0
    # via -r requirements-dev.in
async-timeout==3.0.1
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   pytest
billiard==3.6.4.0
    # via celery
cachetools==5.0.0
    # via google-auth
celery==5.2.2
    # via -r requirements-dev.in
certifi==2021.10.8
    # via
    #   kubernetes
    #   requests
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
click==8.1.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
    #   orchest-cli
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
croniter==1.0.15
    # via -r requirements-dev.in
dnspython==1.16.0
    # via eventlet
eventlet==0.30.2
    # via -r requirements-dev.in
flask==1.1.1
    # via
    #   -r requirements-dev.in
    #   flask-cors
    #   flask-migrate
    #   flask-restx
    #   flask-script
    #   flask-sqlalchemy
    #   orchest-internals
flask-cors==3.0.9
    # via -r requirements-dev.in
flask-migrate==2.5.3
    # via -r requirements-dev.in
flask-restx==0.5.1
    # via -r requirements-dev.in
flask-script==2.0.6
    # via -r requirements-dev.in
flask-sqlalchemy==2.4.1
    # via
    #   -r requirements-dev.in
    #   flask-migrate
google-auth==2.6.2
    # via kubernetes
greenlet==1.1.2
    # via
    #   eventlet
    #   sqlalchemy
idna==2.8
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
itsdangerous==2.0.1
    # via
    #   -r requirements-dev.in
    #   flask
jinja2==3.0.3
    # via
    #   -r requirements-dev.in
    #   flask
jsonschema==4.4.0
    # via flask-restx
kombu==5.2.4
    # via celery
kubernetes==21.7.0
    # via
    #   -r requirements-dev.in
    #   orchest-cli
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.0
    # via requests-oauthlib
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.28
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements-dev.in
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==3.0.7
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.1
    # via pytest-mock
pytest-mock==3.6.1
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via
    #   croniter
    #   kubernetes
python-engineio==3.14.2
    # via
    #   -r requirements-dev.in
    #   python-socketio
python-socketio[client]==4.6.1
    # via -r requirements-dev.in
pytz==2022.1
    # via
    #   apscheduler
    #   celery
    #   flask-restx
    #   tzlocal
pyyaml==6.0
    # via kubernetes
requests==2.22.0
    # via
    #   -r requirements-dev.in
    #   kubernetes
    #   orchest-internals
    #   python-socketio
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via kubernetes
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   -r requirements-dev.in
    #   apscheduler
    #   click-repl
    #   eventlet
    #   flask-cors
    #   flask-restx
    #   google-auth
    #   kubernetes
    #   python-dateutil
    #   python-engineio
    #   python-socketio
    #   sqlalchemy-utils
sqlalchemy==1.4.32
    # via
    #   alembic
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.37.4
    # via -r requirements-dev.in
tomli==2.0.1
    # via pytest
typing-extensions==4.1.1
    # via aiohttp
tzlocal==2.1
    # via apscheduler
urllib3==1.25.11
    # via
    #   kubernetes
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
websocket-client==1.3.2
    # via
    #   kubernetes
    #   python-socketio
werkzeug==2.0.3
    # via
    #   -r requirements-dev.in
    #   flask
    #   flask-restx
yarl==1.7.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
